requirement.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
celery
redis
python-decouple
python-telegram-bot
djangorestframework
djangorestframework-simplejwt